Setting up rswag gem, home api documentation
