+++ Gemfile
@@ -15,6 +15,7 @@ gem 'devise'
 gem 'devise-jwt'
 gem 'faker'
 gem 'fast_jsonapi'
+gem 'rswag'
 
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 1.1'
