+++ config/database.yml
@@ -23,8 +23,8 @@ default: &default
   # host: <%= ENV['POSTGRES_HOST'] %>
   # username: <%= ENV['POSTGRES_USER'] %>
   # password: <%= ENV['POSTGRES_PASSWORD'] %>
-  # username: railuser
-  # password: user
+  username: railuser
+  password: user
 
 development:
   <<: *default
