+++ spec/integration/current_user_spec.rb
