+++ spec/integration/users_spec.rb
